refactor(test) : Context 기반 패키지에 맞춰 테스트 패키지 구조 변경
