+++ src/test/java/kitchenpos/application/infra/FakeKitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.application.infra;
 
 import kitchenpos.infra.KitchenridersClient;
